HADOOP-17039. Change scope of InternalDirOfViewFs and InodeTree to make ViewFileSystem extendable outside common package
